Add files via upload
see code for problem. went back in after i got the code working and simplified it a bit! had an empty array that i added the missing values in between the two given values to, and then concatenated that to the original array, but that was extra so i took it out. woohoo!